+++ src/Loader.php
@@ -1,8 +1,6 @@
 <?php
 namespace RubiconMaps;
 
-use RubiconMaps\Divi\RubiconLocationListModule;
-use RubiconMaps\Divi\RubiconMapModule;
 use RubiconMaps\PostType\Location;
 use RubiconMaps\Taxonomy\LocationCategory;
 use RubiconMaps\Taxonomy\Region;
@@ -36,7 +34,6 @@ class Loader
         add_action('init', [LocationCategory::class, 'register']);
         add_action('init', [Region::class, 'register']);
         add_action('rest_api_init', [LocationsEndpoint::class, 'register_routes']);
-        add_action('et_builder_ready', [self::class, 'register_divi_modules']);
 
         if (is_admin()) {
             SettingsPage::init();
@@ -46,19 +43,5 @@ class Loader
 
         new ListShortcode();
         new MapShortcode();
-    }
-
-    public static function register_divi_modules(): void
-    {
-        if (\RubiconMaps\Support\Plugin::isDivi5Enabled()) {
-            return;
-        }
-
-        if (!class_exists('\ET_Builder_Module')) {
-            return;
-        }
-
-        new RubiconMapModule();
-        new RubiconLocationListModule();
     }
 }